refactor: apply dto path change
