+++ springboot/src/main/kotlin/com/goliath/emojihub/springboot/controller/UserController.kt
@@ -1,8 +1,8 @@
 package com.goliath.emojihub.springboot.controller
 
-import com.goliath.emojihub.springboot.dto.LoginRequest
-import com.goliath.emojihub.springboot.dto.SignUpRequest
-import com.goliath.emojihub.springboot.dto.UserDto
+import com.goliath.emojihub.springboot.dto.user.LoginRequest
+import com.goliath.emojihub.springboot.dto.user.SignUpRequest
+import com.goliath.emojihub.springboot.dto.user.UserDto
 import com.goliath.emojihub.springboot.service.UserService
 import org.springframework.http.HttpStatus
 import org.springframework.http.ResponseEntity
