+++ springboot/src/main/kotlin/com/goliath/emojihub/springboot/service/UserService.kt
@@ -4,9 +4,9 @@ import com.goliath.emojihub.springboot.common.CustomHttp401
 import com.goliath.emojihub.springboot.common.CustomHttp404
 import com.goliath.emojihub.springboot.common.CustomHttp409
 import com.goliath.emojihub.springboot.dao.UserDao
-import com.goliath.emojihub.springboot.dto.LoginRequest
-import com.goliath.emojihub.springboot.dto.SignUpRequest
-import com.goliath.emojihub.springboot.dto.UserDto
+import com.goliath.emojihub.springboot.dto.user.LoginRequest
+import com.goliath.emojihub.springboot.dto.user.SignUpRequest
+import com.goliath.emojihub.springboot.dto.user.UserDto
 import org.springframework.stereotype.Service
 
 interface UserService {
